vayu: overlay: Ignore Phone Services in location indicator
AOSP includes a list of services that are whitelisted and do not show
the location indicator on the status bar.
We need to remove the indicator caused by the Phone Services on 12L, so
use this trick to get this working. It may be overwritten by Gapps, but the
only downside is that in this case the location indicator will be back
on the status bar (if enabled).

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>